code improvements
--restore feature calls correct git command and shows status if nothing
is in the staging area.

Dotfiles.short_status is back to a @cached_property because it's
extensively used and expensive/needless to recompute on each use. The
only reason I ever made it a @property was for testing purposes because
I wanted to programatically move files from one state and confirm the
new state would report as an empty list, []. Got around this by doing
the programatic changes then making a new mydot.Dotfiles() object and
confirming the state being tested now returns an empty list.