extracted .discard_changes and .restore
Continung to move all I can out of the Repository class because it had
far too many responsibilties before. Goal is to get it focused down into
a class that can report about the state of the repository and then use
Actions to make changes.

In the longer run I can see functionality that would take 2 Repository
objects and let me do things like 'follow' friends git repos and compare
their files to mine or even import files/folders from their configs into
my own. The better this tool can weave together multiple sources of
configuration the easier it will be for the community to share and learn
from one another. I can even see an import where I follow someone's
neovim config and tell the program where it's root is and be able to
compare init.vim files even if theirs is at the root of their repo but
mine is in ~/.config/nvim/. By setting the fellow nvim to be rooted at
~/.config/nvim the code can know my ~/.config/nvim/init.vim is the same
file and their init.vim